Move background render utils to utils
Move background render utils to utils
